+++ src/Posseth.Toyota.Client/Models/ToyotaApiResponse.cs
@@ -1,8 +1,0 @@
-namespace Posseth.Toyota.Client.Models
-{
-    public record ToyotaApiResponse(
-        string? Status,
-        string? Message
-        // Voeg hier extra properties toe afhankelijk van de API response
-    );
-}